Assignment2/requirements.txt
msgpack
pyyaml
matplotlib
tornado
